requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
SQLAlchemy<2.0
python-dotenv==1.0.0
PyJWT==2.6.0
python-multipart==0.0.6
pytest==7.2.0
httpx==0.24.1
